jssrc2cpg: fixed member creation for classes and interfaces
Fixes: https://github.com/joernio/joern/issues/2380
